feat: refactor of example multi chat conversation
