+++ examples/06-multi-agent-conversation/memory.py
@@ -1,0 +1,2 @@
+
+messages_history_memory = []
